requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-login
werkzeug
requests
scikit-learn
numpy
pandas
joblib
pyserial
